Unhard coded potential variables
Took all hard-coded values in chart.js out and put them in the namespace
marriage. This should be built and init in data.js. For the time being
this is fine but we'll have to see what polymer offers when the time
comes. Also, not all names and queries have been written, only
implemented.